+++ docs/rules/0216/index.md
@@ -1,5 +1,5 @@
 ---
-aip_listing: 216
+aep_listing: 216
 permalink: /216/
 redirect_from:
   - /0216/
@@ -7,4 +7,4 @@ redirect_from:
 
 # States
 
-{% include linter-aip-listing.md aip=216 %}
+{% include linter-aep-listing.md aep=216 %}
