+++ docs/rules/0216/state-field-output-only.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 216
+  aep: 216
   name: [core, '0216', state-field-output-only]
   summary: Lifecycle State enum fields should be marked OUTPUT_ONLY
 permalink: /216/state-field-output-only
@@ -11,7 +11,7 @@ redirect_from:
 # States
 
 This rule requires that all lifecycle state fields whose names end with `State`
-are marked as `OUTPUT_ONLY`, as mandated in [AIP-216][].
+are marked as `OUTPUT_ONLY`, as mandated in [AEP-216][].
 
 ## Details
 
@@ -59,7 +59,7 @@ State state = 1 [(google.api.field_behavior) = OUTPUT_ONLY];
 ## Disabling
 
 If you need to violate this rule, use a leading comment above the field
-Remember to also include an [aip.dev/not-precedent][] comment explaining why.
+Remember to also include an [aep.dev/not-precedent][] comment explaining why.
 
 ```proto
 enum BookState {
@@ -67,12 +67,12 @@ enum BookState {
 }
 
 // (-- api-linter: core::0216::state-field-output-only=disabled
-//     aip.dev/not-precedent: We need to do this because reasons. --)
+//     aep.dev/not-precedent: We need to do this because reasons. --)
 BookState state = 1;
 ```
 
 If you need to violate this rule for an entire file, place the comment at the
 top of the file.
 
-[aip-216]: https://aip.dev/216
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-216]: https://aep.dev/216
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
