+++ rules/aep0216/synonyms.go
@@ -12,36 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0216
+package aep0216
 
 import (
-	"fmt"
 	"strings"
 
-	"github.com/googleapis/api-linter/lint"
+	"github.com/aep-dev/api-linter/lint"
+	"github.com/aep-dev/api-linter/locations"
 	"github.com/jhump/protoreflect/desc"
 )
 
-var nesting = &lint.EnumRule{
-	Name: lint.NewRuleName(216, "nesting"),
-	OnlyIf: func(e *desc.EnumDescriptor) bool {
-		return strings.HasSuffix(e.GetName(), "State") && e.GetName() != "State"
-	},
+var synonyms = &lint.EnumRule{
+	Name: lint.NewRuleName(216, "synonyms"),
 	LintEnum: func(e *desc.EnumDescriptor) []lint.Problem {
-		messageName := strings.TrimSuffix(e.GetName(), "State")
-		fqMessageName := messageName
-		file := e.GetFile()
-		if pkg := file.GetPackage(); pkg != "" {
-			fqMessageName = pkg + "." + messageName
-		}
-		if file.FindMessage(fqMessageName) != nil {
+		if strings.HasSuffix(e.GetName(), "Status") {
 			return []lint.Problem{{
-				Message: fmt.Sprintf(
-					"Nest %q within %q, and name it `State`.",
-					e.GetName(),
-					messageName,
-				),
+				Message:    `Prefer "State" over "Status" for lifecycle state enums.`,
+				Suggestion: strings.Replace(e.GetName(), "Status", "State", 1),
 				Descriptor: e,
+				Location:   locations.DescriptorName(e),
 			}}
 		}
 		return nil
